client/requirements.txt
opencv-python
requests
pyyaml
numpy
sounddevice
scipy